requirements.txt
pyserial
pyserial-asyncio
pynmea2
pynmeagps
bleak
